[Phase 2][Issue-442] feat: add foundation/i18n_foundation and migrate message catalog
